+++ payzen/translations/fr.php
@@ -485,6 +485,7 @@ $_MODULE['<{payzen}prestashop>payzen_dd367f5fbbdba105ab71cdcf4f7cf366'] = 'Le mo
 $_MODULE['<{payzen}prestashop>payzen_8e83bfe58eb12afb591a7f99f01e6d35'] = 'Il est impossible de rembourser une transaction impayée.';
 $_MODULE['<{payzen}prestashop>payzen_d67793cb999383fca646b7954ab67137'] = 'Une erreur est survenue lors du remboursement de la commande #%1$s. Veuillez consulter les logs de PayZen pour plus de détails.';
 $_MODULE['<{payzen}prestashop>payzen_0ad207a47022c5ba4d99d915d39a06a7'] = 'Impossible de rembourser la commande #%1$s.';
+$_MODULE['<{payzen}prestashop>payzen_7f1405a4460380a57ee6e8cd0d049ae3'] = 'Le remboursement de paiement fractionné n\'est pas supporté. Pensez à apporter les modifications nécessaires dans le Back Office %1$s.';
 $_MODULE['<{payzen}prestashop>payment_std_rest_4c561c57168fbd2c6a31ad79170183cf'] = 'Cliquez sur le bouton « Payer » pour saisir les informations sur le paiement en mode popin';
 $_MODULE['<{payzen}prestashop>back_office_40844e29d3d1c9aca36a6c15e702808a'] = 'Ajouter des moyens de paiement';
 $_MODULE['<{payzen}prestashop>back_office_004bf6c9a40003140292e97330236c53'] = 'Action';